+++ packages/salesforcedx-vscode-core/src/commands/index.ts
@@ -37,6 +37,7 @@ export {
   forceLightningEventCreate,
   forceLightningInterfaceCreate,
   forceLightningLwcCreate,
+  forceLightningLwcTestCreate,
   forceVisualforceComponentCreate,
   forceVisualforcePageCreate
 } from './templates';
@@ -44,6 +45,7 @@ export { forceDebuggerStop } from './forceDebuggerStop';
 export { forceConfigList } from './forceConfigList';
 export { forceAliasList } from './forceAliasList';
 export { forceOrgDisplay } from './forceOrgDisplay';
+export { forcePackageInstall } from './forcePackageInstall';
 export {
   forceSfdxProjectCreate,
   forceProjectWithManifestCreate