+++ packages/salesforcedx-vscode-core/src/notifications/notificationService.ts
@@ -84,13 +84,12 @@ export class NotificationService {
     cancellationToken?: vscode.CancellationToken
   ) {
     observable.subscribe(async data => {
-      if (data !== undefined && data.toString() === '0') {
+      if (data !== undefined && String(data) === '0') {
         await this.showSuccessfulExecution(executionName);
-      } else {
+      } else if (data !== null) {
         this.showFailedExecution(executionName);
       }
     });
-
     if (cancellationToken) {
       cancellationToken.onCancellationRequested(() => {
         this.showCanceledExecution(executionName);