+++ packages/salesforcedx-vscode-lwc/src/index.ts
@@ -34,6 +34,7 @@ import {
   activateLwcTestSupport,
   shouldActivateLwcTestSupport
 } from './testSupport';
+import { WorkspaceUtils } from './util/workspaceUtils';
 
 // See https://github.com/Microsoft/vscode-languageserver-node/issues/105
 export function code2ProtocolConverter(value: Uri) {
@@ -126,6 +127,9 @@ export async function activate(context: ExtensionContext) {
     activateLwcTestSupport(context, workspaceType);
   }
 
+  // Initialize utils for user settings
+  WorkspaceUtils.instance.init(context);
+
   // Notify telemetry that our extension is now active
   telemetryService.sendExtensionActivationEvent(extensionHRStart).catch();
 }
@@ -144,7 +148,7 @@ function getActivationMode(): string {
 }
 
 function registerCommands(
-  extensionContext: vscode.ExtensionContext
+  _extensionContext: vscode.ExtensionContext
 ): vscode.Disposable {
   return vscode.Disposable.from(
     vscode.commands.registerCommand(